src/copilot_flow/requirements.txt
openai
azure-identity
azure-search-documents==11.4.0
promptflow[azure]==1.11.0
promptflow-tracing==1.11.0
promptflow-tools==1.11.0
jinja2
aiohttp
python-dotenv